chore: remove unnecessary extra files
